requirements8.txt
absl-py==2.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
Flask==3.0.3
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
gradio==5.33.0
gradio_client==1.10.2
groovy==0.1.2
grpcio==1.72.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==10.2.0
protobuf==5.29.5
py4j==0.10.9.7
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.15.2
tomlkit==0.13.3
torch==2.2.0
tqdm==4.67.1
transformers==4.38.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
